feat: infer weight checkpoint layout from LoRA
Weight checkpoints have no knobs left: full runs save sharded
safetensors through the rank-parallel path, LoRA runs save only the
adapter (peft layout, at the step dir root) plus config and tokenizer.
save_adapter_separately is removed, and with it the merged-LoRA
gather-to-master path (gather_weights_on_master, save_to_path,
clean_lora_state_dict). The benchmark script keeps a vendored copy of
the master gather as its baseline.
