+++ configs/ci/integration/reverse-text-sft-lora/start.toml
@@ -2,9 +2,6 @@ max_steps = 5
 
 [ckpt]
 
-[ckpt.weights]
-save_adapter_separately = true
-
 [model]
 name = "PrimeIntellect/Qwen3-0.6B"
