+++ configs/debug/algo/opd_lora.toml
@@ -70,9 +70,6 @@ lr = 1e-4
 [trainer.model.lora]
 rank = 8
 
-[trainer.ckpt.weights]
-save_adapter_separately = true
-
 [ckpt]
 
 [inference]
