+++ examples/vlm_sft_moe/sft.toml
@@ -47,9 +47,6 @@ lr = 1e-5
 
 [ckpt]
 interval = 200
-[ckpt.weights]
-save_adapter_separately = true
-
 [deployment]
 type = "single_node"
 num_train_gpus = 8
